pyproject.toml
[project]
name = "slayer"
version = "0.1.0"
description = "Company Research Tool for Job Seekers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
company-research = "company_research.cli:main"
